fix: ・話数に小数(第4.5話など)が来ると録画同期がパースエラーで停止する問題を修正 ・小数の話数を保存できるようDBバージョンを15へ更新 ・局ロゴの取得をIOスレッドで行うよう修正 (メインスレッドでの通信によりロゴが表示されなかった) ・局ロゴの同時取得数を制限し、チャンネル数が多い環境でサーバーへ負荷が集中しないよう調整

+++ app/src/main/java/com/beeregg2001/komorebi/data/local/entity/RecordedProgramEntity.kt
@@ -41,7 +41,7 @@ data class RecordedProgramEntity(
     @ColumnInfo(name = "direct_thumbnail_url") val directThumbnailUrl: String? = null,
     @ColumnInfo(name = "api_thumbnail_url") val apiThumbnailUrl: String? = null,
     @ColumnInfo(name = "series_id") val seriesId: Int? = null,
-    @ColumnInfo(name = "episode_number") val episodeNumber: Int? = null,
+    @ColumnInfo(name = "episode_number") val episodeNumber: Double? = null,
     @ColumnInfo(name = "episode_label") val episodeLabel: String? = null,
     @ColumnInfo(name = "episode_title") val episodeTitle: String? = null,
     @ColumnInfo(name = "air_type") val airType: String? = null
